Remove redundant casts.

+++ EDFREQ/src/edfreq/region/DoubleSideBandRangeBar.java
@@ -147,7 +147,7 @@ public class DoubleSideBandRangeBar extends JPanel implements Observer
 
 		for( int i = 0 ; i < _xRanges.size() ; i++ )
 		{
-			int value = ( int )( _xRanges.elementAt( i ).intValue() * zoomChange ) ;
+			int value = ( int )( _xRanges.elementAt( i ) * zoomChange ) ;
 			_xRanges.setElementAt( new Integer( value ) , i ) ;
 		}
 
@@ -171,8 +171,8 @@ public class DoubleSideBandRangeBar extends JPanel implements Observer
 		// The RangeBar range refers to the backend bandwidth.
 		for( int i = 0 ; i < _xRanges.size() ; i += 2 )
 		{
-			_lowerRangeBar.setRange( _feBandWidth - _xRanges.get( i + 1 ).intValue() , _feBandWidth - _xRanges.get( i ).intValue() , i / 2 ) ;
-			_upperRangeBar.setRange( _xRanges.get( i ).intValue() , _xRanges.get( i + 1 ).intValue() , i / 2 ) ;
+			_lowerRangeBar.setRange( _feBandWidth - _xRanges.get( i + 1 ) , _feBandWidth - _xRanges.get( i ) , i / 2 ) ;
+			_upperRangeBar.setRange( _xRanges.get( i ) , _xRanges.get( i + 1 ) , i / 2 ) ;
 		}
 	}
